requirements.txt
Cheetah==2.4.4
Django==1.10.4
Landscape-Client==14.12
Markdown==2.6.6
Mezzanine==4.2.2
PAM==0.4.2
Pillow==3.4.2
PyYAML==3.10
Twisted-Core==13.2.0
Unidecode==0.04.19
Wand==0.4.4
apt-xapian-index==0.45
argparse==1.2.1
bbcode==1.0.22
beautifulsoup4==4.5.1
bleach==1.5.0
chardet==2.3.0
cloud-init==0.7.5
cloudinary==1.5.0
colorama==0.2.5
configobj==4.7.2
construct==2.5.2
coverage==4.0.2
django-admin-sortable==2.0.10
django-annoying==0.9.0
django-appconf==1.0.2
django-compressor==2.0
django-contrib-comments==1.7.3
django-datetime-widget==0.9.3
django-debug-toolbar==1.6
django-embed-video==1.1.0
django-extensions==1.6.7
django-floppyforms==1.7.0
django-friendship==1.3.1
django-mailer==1.1
django-notification==1.3.3
django-request==1.5.1
django-scheduler==0.8
django-taggit==0.20.2
django-user-accounts==1.3.1
enum34==1.1.6
filebrowser-safe==0.4.6
future==0.16.0
grappelli-safe==0.4.5
gunicorn==19.4.1
html5lib==0.9999999
icalendar==3.9.2
ipaddress==1.0.14
jsonpatch==1.3
jsonpointer==1.0
lockfile==0.12.2
oauth==1.0.1
oauthlib==2.0.1
opentok==2.4.1
pefile==2016.3.28
pep8==1.7.0
pluggy==0.3.1
prettytable==0.7.2
psycopg2==2.6.1
py==1.4.31
pyOpenSSL==0.13
pycurl==7.19.3
pyflakes==1.1.0
pygeoip==0.3.2
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-dateutil==2.6.0
python-debian==0.1.21-nmu2ubuntu2
python-ptrace==0.9
pytz==2016.10
rcssmin==1.0.6
requests==2.12.4
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.2
ssh-import-id==3.21
tox==2.3.1
tzlocal==1.3
urllib3==1.19.1
uuid==1.30
virtualenv==15.0.1
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
